+++ integrations/fullstory/lib/index.js
@@ -17,7 +17,9 @@ var integration = require('@segment/analytics.js-integration');
 var FullStory = (module.exports = integration('FullStory')
   .option('org', '')
   .option('debug', false)
-  .tag('<script src="https://www.fullstory.com/s/fs.js"></script>'));
+  .tag(
+    '<script async src="https://www.fullstory.com/s/fs.js" crossorigin="anonymous"></script>'
+  ));
 
 /**
  * The ApiSource string.
@@ -37,11 +39,13 @@ FullStory.prototype.initialize = function() {
 
   /* eslint-disable */
   /* istanbul ignore next */
+  /* The snippet below differs slightly from the snippet available on fullstory.com because fs.js is already loaded above*/
   (function(m,n,e,t,l,o,g,y){
     if (e in m) {if(m.console && m.console.log) { m.console.log('FullStory namespace conflict. Please set window["_fs_namespace"].');} return;}
     g=m[e]=function(a,b,s){g.q?g.q.push([a,b,s]):g._api(a,b,s);};g.q=[];
     g.identify=function(i,v,s){g(l,{uid:i},s);if(v)g(l,v,s)};g.setUserVars=function(v,s){g(l,v,s)};g.event=function(i,v,s){g('event',{n:i,p:v},s)};
     g.shutdown=function(){g("rec",!1)};g.restart=function(){g("rec",!0)};
+    g.log = function(a,b) { g("log", [a,b]) };
     g.consent=function(a){g("consent",!arguments.length||a)};
     g.identifyAccount=function(i,v){o='account';v=v||{};v.acctId=i;g(o,v)};
     g.clearUserCookie=function(){};
@@ -73,12 +77,13 @@ FullStory.prototype.identify = function(identify) {
 
   var newTraits = foldl(
     function(results, value, key) {
+      var rs = results;
       if (key !== 'id') {
-        results[
+        rs[
           key === 'displayName' || key === 'email' ? key : camelCaseField(key)
         ] = value;
       }
-      return results;
+      return rs;
     },
     {},
     traits