+++ integrations/amplitude/HISTORY.md
@@ -1,3 +1,16 @@
+# 3.3.0 / 2019-10-09
+
+- Add support for versionName setting, this will allow user to specify version.
+- Add support for trackProductsOnce setting, this will track all products at once.
+
+# 3.2.0 / 2019-08-27
+
+- Add support for unsetParamsReferrerOnNewSession setting
+
+# 3.1.0 / 2019-08-09
+
+- Add mapping of non-property fields to event_props.
+
 # 3.0.0 / 2019-06-10
 
 - Update Amplitude SDK to v5.2.2